requirements.txt
langchain
langchain-community
langchainhub
langchain-openai
langchain-chroma
tiktoken
sentence-transformers
pypdf
chromadb
fastapi
uvicorn
pydantic
python-telegram-bot>=20.0
asyncio
httpx
aiofiles
google-cloud-storage>=2.0.0
google-auth>=2.0.0
boto3